Move json functions in separate file.
