+++ runtime/runtime.cmake
@@ -23,6 +23,7 @@ prepend(KPHP_RUNTIME_SOURCES ${BASE_DIR}/runtime/
         instance_cache.cpp
         inter-process-mutex.cpp
         interface.cpp
+        json-functions.cpp
         kphp-backtrace.cpp
         mail.cpp
         math_functions.cpp
