+++ runtime/vkext.cpp
@@ -616,10 +616,6 @@ string f$vk_flex(const string &name, const string &case_name, int64_t sex, const
   return string(buff);
 }
 
-Optional<string> f$vk_json_encode(const mixed &v) {
-  return f$json_encode(v, 0, true);
-}
-
 string f$vk_whitespace_pack(const string &str, bool html_opt) {
   const char *text = str.c_str();
   int level = 0;
